LibWeb: Move backing store allocation into manager
CompositorThread still owned the backing-store surfaces and allocation
helpers after BackingStoreManager moved into the compositor. That left
the thread class responsible for low-level surface creation details even
though the manager now decides when backing stores need to change.
